put the saga in chat, not window
