+++ shared/actions/window.js
@@ -1,36 +1,8 @@
 // @flow
 import * as Constants from '../constants/window'
-import {safeTakeEvery} from '../util/saga'
-import {put, select} from 'redux-saga/effects'
-import {selectConversation} from '../actions/chat'
-import {chatTab} from '../constants/tabs'
 
-import type {ChangedFocus} from '../constants/window'
-import type {SagaGenerator} from '../constants/types/saga'
-import type {TypedState} from '../constants/reducer'
+import type {ChangedFocused} from '../constants/window'
 
-export function changedFocus (focus: boolean): ChangedFocus {
+export function changedFocus (focus: boolean): ChangedFocused {
   return {type: Constants.changedFocus, payload: focus}
 }
-
-function * _changedFocusSaga (action: ChangedFocus): SagaGenerator<any, any> {
-  const focused = action.payload
-  // Reselect the current Chat conversation, to give badging a chance to
-  // update based on the focus change.
-  const selectedSelector = (state: TypedState) => state.chat.get('selectedConversation')
-  const conversationIDKey = yield select(selectedSelector)
-  const routeSelector = (state: TypedState) => state.routeTree.get('routeState').get('selected')
-  const selectedTab = yield select(routeSelector)
-  const chatTabSelected = (selectedTab === chatTab)
-  if (conversationIDKey && focused && chatTabSelected) {
-    yield put(selectConversation(conversationIDKey, true))
-  }
-}
-
-function * windowSaga (): SagaGenerator<any, any> {
-  yield [
-    safeTakeEvery(Constants.changedFocus, _changedFocusSaga),
-  ]
-}
-
-export default windowSaga
