+++ shared/store/configure-sagas.js
@@ -11,7 +11,6 @@ import planBillingSaga from '../actions/plan-billing'
 import profileSaga from '../actions/profile'
 import settingsSaga from '../actions/settings'
 import pushSaga from '../actions/push'
-import windowSaga from '../actions/window'
 import {call} from 'redux-saga/effects'
 
 import type {SagaGenerator} from '../constants/types/saga'
@@ -29,7 +28,6 @@ function * mainSaga (): SagaGenerator<any, any> {
     call(profileSaga),
     call(pushSaga),
     call(settingsSaga),
-    call(windowSaga),
   ]
 }
